1.0.1
Launch blacklisted Sandswept Delta Construct because they always error on death (not my fault but it breaks the game with launch)

Added CanBeOverridden to FormStateBase that determines whether a form can be cancelled by transforming into another form

I need to network Launch values for when you restart a launch mid-flight

Moved checks for launching enemies into their own static methods. TargetCanBeLaunched is whether a target should be able to have a launch started on it, regardless of situation.
AttackCanLaunch is whether an attack is able to Launch given a target and damageinfo. This is what checks for stuff like mass, proc coefficient, etc.